Parsing fragment supports arbitrary keys
This switches parsing the fragment (string following '#' at the end of a requirement) from looking for 'egg' at the beginning of the fragment to allowing any key value pair. These can be extracted as needed - right now only the 'egg' key is used to get the name of the package.

Tests include URI with a single fragment, and multiple fragments.