chore(deps): remove obsolete `types-setuptools`
Before https://github.com/madpah/requirements-parser/pull/105,
`pkg_resources` was implicitly present with `setuptools`, so it was
useful to have `types-setuptools` to check typing. But now that the
usage is gone, the typing dependency should not be needed. It probably
should have never been a direct dependency anyway, but only a dev
dependency, as AFAIK it was only useful locally to verify typing, not
when installing the library for consumers.

Signed-off-by: Mathieu Kniewallner <mathieu.kniewallner@gmail.com>